requirements.txt
streamlit>=1.35.0
supabase>=2.4.0
pandas>=2.0.0
plotly>=5.18.0
folium>=0.16.0
streamlit-folium>=0.20.0
python-dotenv>=1.0.0
numpy>=1.26.0
requests>=2.31.0
